src/main/java/ru/netology/movieManager.java
package ru.netology;

public class movieManager {
    private final int limit;
    private String[] movies = new String[0];

    public movieManager() {
        this.limit = 5;
    }

    public movieManager(int limit) {
        this.limit = limit;
    }

    public void add(String movie) {
        String[] tmp = new String[movies.length + 1];
        for (int i = 0; i < movies.length; i++) {
            tmp[i] = movies[i];
        }
        tmp[tmp.length - 1] = movie;
        this.movies = tmp;
    }

    public String[] FindAll() {
        return movies;
    }

    public String[] FindLast() {
        int resultLenght;
        if (movies.length < limit) {
            resultLenght = movies.length;
        } else {
            resultLenght = limit;
        }
        String[] ans = new String[resultLenght];
        for (int i = 0; i < ans.length; i++) {
            ans[i] = movies[movies.length - 1 - i];
        }
        return ans;
    }
}